Fix icons
- Unify getNewSteps.
- Default to no icon if no step identified.
- Make APB icon more permissive so that it includes 223 block in Back.
