+++ todo.txt
@@ -13,15 +13,11 @@ So like AUF + trigger_1 + y + AUF + trigger_2?
 Trigger_1 = any RU or LU
 Trigger_2 = any RU, LU, or FU
 
-Use direct detection from block object to see if FB SB are solved.
-They might be part of one block and improperly bounded.
-
 FMC icons. DR/HTR? What even.
 
 Rank OLL and PLL suggestions without algorithm, just hard code it. Use reco.nz ranks.
 
 A completely intuitive f2l guide
 
-Randomly generated titles for hyperlink upon Ctrl+S preview copy
-
-If both L-Block and R-Square present in currentSteps, we should always show block, assuming 
+Comments in previews, such as // OCLS break (writes: // OCL S)
+Possibly just an issue where moving text around force adds break between them
